updated codebase structure
